Jump back in the background on pause.
This moves the logic from onPlayBackResumed to onPlayBackPaused.  A timer is
started there to jump back at the desidered (calculated) percentage (there's no
SeekTime built-in function) and it is cancelled when manually resumed or
playback is stopped (the call to SeekPercentage is ignored if nothing is
playing, but this is needed for e.g. playlists).

On a device like a Raspberry Pi the previous logic randomly caused the following
behavior:
  * user resumes playback;
  * video plays for a second or two;
  * it finally jumps back.
Although I'm not sure I believe this behaves as intended, as Python threads have
a lower priority than player's ones.

Additionally, in case of a high bit rate video it may need a couple seconds to
be cached.  By rewinding while still paused it is cached "in the background", so
the user hits the resume button and it just plays.